- Fixed a battlescape startup crash. - Fixed some resources being loaded where they shouldn't.

+++ src/Ruleset/MapDataSet.cpp
@@ -80,7 +80,7 @@ std::string MapDataSet::getName() const
 * Gets the MapDataSet size.
 * @return size in number of records.
 */
-size_t MapDataSet::getSize() const
+int MapDataSet::getSize() const
 {
 	return _objects.size();
 }
